app/requirements.txt
boto3~=1.21
Pillow~=9.0